adds Cluster resource management
Adds the Cluster as the first managed resource. The MSK Cluster created
in the e2e test is the smallest possible cluster (2 nodes, small
instance types) but still takes on average around 25-35 minutes to
create. This means the e2e test run is very long, but there's nothing
really we can do about that for now.

Issue aws-controllers-k8s/community#348

Signed-off-by: Jay Pipes <jaypipes@gmail.com>