updating the olm config (#2)
Issue #, if available:
N/A
Description of changes:
Updating the OLM description. There are a few different MSK/Kafka description pages, none of them offer the same information, so I wasn't sure what to pick for branding. Attached is a picture of the overall look/feel, let me know if something else needs to change branding wise.

![image](https://user-images.githubusercontent.com/88156657/208526770-1eece493-0343-4190-a3dc-88c632fa4ae6.png)


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

Signed-off-by: Adam D. Cornett <adc@redhat.com>