Add bootstrap broker string fields to Cluster status (#41)
this commit adds several new fields to the `Cluster` Status
struct to store the bootstrap broker string information for
Amazon MSK clusterss. These fields include:

- `BootstrapBrokerString`
- `BootstrapBrokerStringPublicSASLIAM`
- `BootstrapBrokerStringPublicSASLSCRAM`
- `BootstrapBrokerStringPublicTLS`
- `BootstrapBrokerStringSASLIAM`
- `BootstrapBrokerStringSASLSCRAM`
- `BootstrapBrokerStringTLS`
- `BootstrapBrokerStringVPCConnectivitySASLIAM`
- `BootstrapBrokerStringVPCConnectivitySASLSCRAM`
- `BootstrapBrokerStringVPCConnectivityTLS`

The `setResourceAdditionalFields` function is added to the
resource manager to fetch and set these fields from the AWS
API by calls the `GetBootstrapBrokers` API and populates
the corresponding fields in the ClusterStatus.

additionally, we're adding a `customPreCompar`e hook function
to properly set default values for various Cluster spec fields
before comparing resources during the delta calculation - this
is fixing an observed bug that was impact the status fields

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.