Add update operations for VolumeSize, InstanceType, and BrokerCount (#55)
Issue [#2249](https://github.com/aws-controllers-k8s/community/issues/2249)

Description of changes:
Also added reconcile before deletion attempt if cluster is 
not active. Before, if we tried deleting a cluster that was 
not active, the controller would set it to terminal, and does 
not reconcile.

Another change being added is a field in the Cluster status
called ClusterVersion. This field is returned by the sdkFind 
operation, and is crutial to make updates to VolumeSize,
InstanceType, and BrokerCount.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.