Add sdkFind hook to retrieve ServerProperties, a required field (#60)
Description of changes:
We need to retrieve this field to allow this resource to be adopted. 
Adoption does not go through if a required field is missing from the 
spec, and this change fixes that

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.